fix: form validation
